driver ueye: improve reliability when changing settings and start/stop
The IDS ueye API already handles changing the exposure time or gain while the
acqusition is running. Especially, even when stopping/starting the acquisition,
the first image acquired could be with the previous settings.
=> keep things simple, and immediately change the settings when the VAs are changed.

When stopping/starting quickly, the camera could sometimes become irresponsive.
=> refactor the acquisition thread to be like picoquant, ie run constantly and
communicate via a queue. That means that only one thread is talking to the API/camera
and seems to help a lot in stability.

Also make sure the black level is set to auto at init.