acq/gui: Only disable the AutoBC of the stream if the detector supports brightness/contrast control
For the Phenom, which returns data already scaled on 8bpp, we had simplified
the logic for the AutoBC, so that it's disabled by defaulted.

It turns out that there are other detectors that can return 8bpp data (eg, ueye CCDs),
but the data is not scaled. So in this case it doesn't make sense to disable AutoBC.

=> only disable if the detector is 8bpp AND supports scaling the data (via the brightness/contrast VAs).

Note: that will change the behaviour for static streams with 8bpp, reverting back to having AutoBC by default.