driver actuator MultiplexActuator: allow to pass a dict for ref_on_init to indicate default init position
With ref_on_init, the axis would be referenced (and stay at 0) at init.
However, on somecase, the reference position is not a good one.
=> allow to indicate a default position to take after referencing.

We need to differentiate between position 0, which is the reference position,
and not moving at all, in case the axis is already referenced at init.
=> allow to pass None (aka null in YAML) to indicate no move should be done.