config odemis-bug-report: do not block if odemis list hw blocks
If the backend is not behaving well (eg, one driver is killed), listing the VAs can either take a really long time
(> 1 min) or could even block forever, instead of taking just a few seconds.

So do not wait for it to finish to show the description window, and automatically stop it after a while.