acq/gui: avoid holding references to ensure garbage recollection
In most cases, all the objects would eventually be recollected, but
only after an other acquisition has been run or a new stream has been
prepared.

Also call a bit more often the garbage collector when it's very likely to be useful:
just after removing streams.

This helps a bit in memory usage of the GUI. There are still issues with the histogram computation and
AR static stream projection (running in a separate thread).