driver andorcam2: if camera has external shutter, allow using it
External shutter is not detectable from the API. The hardware just has an "output",
and there is no way to know whether it's connected to a real shutter or not.
So by default we don't consider it a shutter.
However, if the user specifies shutter_times, then assume it's connected to a real shutter