acq/gui: more tweaking for memory optimisation
After more experimentation, it comes out that the matplotlib triangulation leaks
memory if multiple threads are using it. Most of the time, the memory can be recovered
if a thread that has used it is closed before the next thread starts.

So, when opening a new file, try to get rid of the preivous streams as soon as possible.

Also do some minor clean ups