driver andorshrk: add drives_shutter argument to force external shutter mode for some output ports
This extends the support for shutter on SR193.
If two cameras are connected to the spectrograph, often at most one will drive the shutter,
the other one expects the shutter to be open all the time.
So introduce a new argument to indicate which camera takes care of handling the shutter,
and which camera needs the shutter always opened.

Note: only tested on simulator, not yet on real hardware.