driver phenom: simplify grid scanning mode
For 1 < res <= 128, the scanning needs to be emulated by multiple spot,
as the Phenom doesn't support such scanning.

To do so, the code was using a separate thread. This caused lots of complexity
to synchronised the scanning with the acquisition. In particular, it would
only stop after the end of the whole scan, and could run _in parallel_ to
another acquisition (especially noticeable if 128x128 was selected).
Moreover, in such mode, the fake data of the detector would be sent at a
very high rate not related to the actual scan.

=> simplify by in-lining the scanning in the standard function.

Also simplify a few other places.