gui conf: handle VA configuration with both choices and range
In that case, it should mean to choices, but with the choices limited to the range.

Also add "range_1d" to do the same with the binning_1d_from_2d special wrapper.

In particular, this ensures that the SEM scale choices on the DELPHI don't include 16.
We had already tried, but failed to realise it wasn't handle by the GUI.