acq mngr/stream: don't prepare the stream after starting acquisition
The acquisition manager was first starting the stream, and then preparing it.
It would wait the right amount of time, but could cause "preparation" to run while
the stream is active. => put in the right order.

Also now log such trial to detect such case... until we raise an error.