acq/gui spectrum export: fix raw line spectrum actually exporting RGB
Change the stream methods to acccept "raw" argument, so that they return the _almost_ raw
values (ie, values in the same range and type as the original data).

Also change the conversion function to take the stream as argument, instead of taking
already half-converted data.

Also allow the user to change the AutoBC outliers value for line spectrum.