driver tmcm: support rng argument to indicate the physical range of the axis
It's useful to know the _real_ range of the axis, instead of just having the controller encoding limitation.
So let the user provide that information.

When the axis is not referenced, we double that range, to be sure we allow the axis to go anywhere it
possibly could go. It also ensures that it'd be possible to reach the limit of the axis,
but that's always the case when an axis is not referenced.