acq delphi: fix calibration going out of the optical stage range
The calibration tries to check the point at the four ordinal point, at 4mm.
That's precisely the range of the optical stage. However, there is the margin for
the backlash + the offset for the SEM stage.
=> Don't use the SEM stage to move at the ordinal points, but the optical stage, to be sure to stick within range.
=> Restrict the 4 positions to 3.5mm to always fit within optical stage range.

Also use the first (center) point for Twin stage transformation matrix fitting, so we use 5 instead of 4

Also center the spot on CCD before doing the last 2 calibration steps, to make sure the CL is on the middle of the CCD.

Also improve the delphi manual script:
* Separate the main() call from the calibration function
* Simplify getch()
* Add --log-level argument