gui conf: simplify format_choices() and increase number of HFW choices
Avoid needing too much info from the config, and just guess when to use the
same unit prefix based on the magnitude.

Also increase the number of HFW choices. We would provide every power of 10,
however for the user 10x is huge. => provide 2x, 5x and 10x for each magnitude.

Also make sure the choices are ordered on the combobox.