gui cont microscope: don't hide streams on stage/focus move on the SECOM
On the SECOM, the stage and focus are closed-loop, so their position keep updating
(so slightly).
So even if two stream acquisition are done without any user-requested move,
their position is likely a tiny bit different. That doesn't mean the non-playing
streams should be hidden, otherwise it's impossible to show two streams simultaneously.

=> only hide the streams on stage move on the DELPHI.