driver phenom: fix spot mode sometimes not being applied
To avoid setting the view parameters again and again, they are cached.
However, they are automatically changed when the SEM goes out of SEM mode.
So it could happen (in very rare cases, mostly during calibration) that
the driver think it's already in spot mode, while actually the SEM is
scanning a whole area.