acq align delphi: restrict positions for twin stage calibration
This is a minimal backport from development version.
The range of the optical stage is about -4->4mm, but the origin is not necessarily precisely at the center.
So trying to position the stage at each border can easily go wrong, causing twin stage calibration to fail
or to be wrong.

=> make the position a bit closer from the center.