driver ueye: Handle better hardware errors during acquisition
Quite a few errors can happen during the acquisition, even when running normally.
Especially, USB transfer failure happen just from time to time, and the only thing
needed is to ask for a new frame (=> no need to wait 1s and reset the acquisition).

Also handle when the camera is disconnected/reconnected.