driver semcomedi: don't wait for trigger again on acquisition retrial
If the acquisition fails (rare, but seems to sometimes happen with no data read in time),
the acquisition would automatically reset the device and retry.
However, if the dataflow are synchronised, they will block again waiting for _another_ trigger,
while the client do not expect to have to send another trigger as no data has arrived yet.

=> don't wait for dataflows if retyring acquisition (ie, nfailures > 0)