driver phenom: Changes spot shift handling and add features to measure it properly
The spot shift need to also include the compensation for the resolution and HFW,
as the spot shift is independent from these calibration values, but they affect
the placement of the SEM image, so the spot position is also affected.

Add fovScale as a way to control the Phenom scale, which is used just for computing
the spot shift (will be introduced in a later commit).

Also changes the spot resolution to 256x256 as it now works and allows a faster
scanning of spot grids.

Also rearanges the way the scan parameters are computed, so that it's not
necessary anymore to store two copies of them constantly... hoping they are
really in sync with the Phenom ones.