driver phenom: remove non-standard fovScale and sneek same functionality in scale
Until now, the acquisition code would not care really about the scale, and only pick the resolution.
So using a resolution less than the full possible resolution possible by the
current scale would still acquire at the full FoV, while the conventional behaviour is
to keep the pixel size (as it should be based on the scale), and only acquire at the center.

=> reduce the Phenom scale in such case.
(It should still only be used for calibration)

As previously the scale wouldn't matter, there _might_ be bugs in code which wasn't
setting the scale properly before changing the resolution.