acq align delphi: Simplify AlignAndOffset() and improve focus during hole detection
When looking for the spot in AlignAndOffset(), if it to find the spot, don't retry everything
again with just a slightly different focus starting point. The exhaustive autofocus should
pick the spot, if it's possible to pick it up.

Also run the SEM autofocus slightly away from the expected hole, to avoid trying to focus on
a hole, which is pretty impossible. If the hole detection fails, it's not a big issue,
but the focus is actually really important, to get the e-beam well focused on the sample too
in later steps.

Also do so small improvements on the log messages.