acq align delphi: New FindRingCenter() function for detecting the position of the lens
The position of the lens is critical to place the SEM stage at the right place and be able to
see the CL spot in the CCD. So the value read from LensAlignment must be as precise as possible.
This new function is more ad-hoc and looks for both inner and outter circles of the lens.
It seems to greatly improve the reliability.

Also reorganise the steps to have the fine alignment performed immediately after the twin stage
calibration, as it also uses the CL spot (and the focus can drift).

Also a few more tweaks to improve the reliability of the whole calibration.