acq align: Give some chances to delphi.AlignOffset() to relocate source
In case AlignOffset() fails to find a spot, it'll try to get closer from the source.
The previous simplification removed the CCD settings, which meant the source
was looked for with CCD settings 0.1 s and binning 1,1. By making it 8,8 @ 1s,
we get 160x more signal, which should improve the chance the brightess spot is
really nearer the source.
Also turn on the SEM, as this could help a bit too.

Also force AlignSpot() to use exhaustive AutoFocus, as it's especially designed to
work better with spot.