driver omicronxx: change default excitation bandwidth from 40nm to 20nm
In the dataio drivers, when no bandwidth information is known, the default is 20 nm.
For concistency in Odemis, also use that default in the driver.
The main advantage is that in the GUI, the same bandwidth info will be shown in acquisition mode
and when opening a file. It is still incorrect, but at least, it's concistent.