gui: Force sample holder ejection at the end of calibration
Either the calibration failed, and so the sample holder needs to be ejected,
or it worked, and it should be ejected to force loading the new value (and
using the current sample is most likely useless).
This also ensures the sample is not loaded while the status message is hidden.

Also reorganise a bit the status message code to ensure it's always shown when needed.