gui comp legend: don't let merge slider take space when not shown
It was originally thought to be a good idea that the rest of the legend doesn't move
when the slider shows/hide. However, when the legend is small, and there is no reason
a slider ever appears this is really annoying, because the scale doesn't have enough space.

Also, in practice, it doesn't feel weird that things move a bit when the slider appears.