gui view: Ensure the hardware HFW stays as-is when the SEM view is resized and handle non 1:1 ratio
Extend the viewport mpp<->fov code to support non 1:1 ratio of the SEM image, which is the case for
Hitachi SEMs. That avoids having black bars both horizontally and vertically.
=> change the .fov_va to fov_hw to be able to access the aspect ratio information too.

Also changes the code to absolutely ensure that hardware HFW doesn't change when the canvas is resized.

Also modify a bit the way the mpp is recomputed (when no hardware HFW is present).
Instead of keeping the minimum dimension, which is not working well when resizing switches
the minimum dimension, we keep the physical area constant.

Also attempt to simplify the code that keeps the HFW of the overview view large.