driver omicronxx/powerctrl: add/improve locking
Take a lock on a read-only file first to make sure that if the file is locked,
we don't change anything on the file.

That's particularly useful with the special "Omicron" lasers which are actually Cobolt DPSS,
as these later ones are also connected via FTDI, but will be disturbed if someone else
tries to access them with wrong settings while already one driver is connected.