gui cont tabs: simplify SECOM alignment tab by using local VAs
Now that there are stream local VA, lots of TODOs can be completed.
In particular, we don't need the hardware settings panel, there are just
two streams, with all the settings.

Mainly this avoid having to use tricks for updating the fine alignment dwell time.
It was not even correctly set when going to the tab, while the spot mode is active.
That also allows to separate the optical settings in the acquisition (ei, with light)
from the settings in the alignment tab (ie, with CL spot).

At the same time, make F6 (play/pause) work properly.