gui cont tabs: add CCD temperature display in SPARCv2 alignment tab
In theory, the CCD temperature should always be fine... but if it's not it's convenient
to have a place to see it in the GUI. So put it in the alignment tab, as for the SPARCv1.

Also fix an error happening if trying to display 2x2 view in alignment tab (eg, by pressing F5).
It just shouldn't be possible to do so.