model: add new metadata HW_NOTE
Will allow to store raw information about the hardware state when an acquisition is performed.
Pretty much all the rest of the metadata is to store the physical conditions of an acquisition.

The goal of this metadata is to store whichever information from a hardware settings perspective.
It should contain all known state of all the hardware involved in the acquisition.