gui main: create GUI logger a bit earlier
This allows more messages to appear in the log panel. In particular, issues with loading the plugins
will now appear in the GUI.

There is still a bug: the log level is not properly set for the button (so the button will not turn red/orange
if there are errors). This is likely related to the ordering of CallAfter().

Also clean-up on_debug, and make it reset the level also when closing the panel, so
that the button doesn't turn red/orange when closing the panel.