gui tabs: Allow zoom/drag on SPARCv2 lens alignment
Long requested feature from users, as the spot can get really small, as well as the alignment precision.

It basically just requires to use a standard view/canvas instead of a special AR canvas.

Also remove possibility to change the focus, as the recommended usage is to just use auto-focus (which
is apparently very reliable on these conditions).