odemis-start: improve logging
Display popup that the backend is starting before calling MODEL_SELECTOR (which can take ~5s).

In case of error in the MODEL_SELECTOR, show a message box, possibly with the message from the model selector.

In case of error when starting the backend, eg the filename for the YAML file is wrong, show a message box.

Also show version number of Odemis in the window.