acq path: use separate focus in chamber-view
In chamber-view, the focus needed is (slightly) different.
Until now, the user could change the focus in chamber-view, but that meant that he had to
check the focus again afterwards to ensure good spectrum/AR acquisition.
This is easy to forget.

=> Store/restore focus value when entering/exiting chamber-view.

Also refactor the test cases, to make it easier to read.