acq stream: fix data range computation when max value is a power of 2
When the max value was excatly a power of 2, it would round down the range
instead of rounding up, leading to the maximum value not being displayed.
This is particularly noticeable on counting PMT acquisitions, when very small
values are counted (eg: up to 2 or 4). In that case, the maximum value would
not be shown.

=> add 1 to the diff