acq stream: re-allow to update the drange based on data content in live streams
For detectors with very large data range, it's really useful to compute the drange
based on the actual data. This was disabled because it was too costly in terms of computation.

=> Move the drange update to histogram (thread), and detect when the drange
is only based on the detector info, so that a large drange can be shrink back
in some cases.