gui conf: change spot_shift to scale_shift in calibration file
As this calibration value has changed for the new release, it's preferable to use
a different configuration variable name.
This way, the new version of Odemis will not use old value, which could bring a
completely wrong spot shift. Which is especially bad, as the fallback value seems
to work pretty well on most systems.
In addition, this means that rolling back to an older system will still allow to
keep the "good" old calibration value (if it was correct in the first place).