driver andorcam2: fix detecting EM gain support
It needs to use a hasSetFunction() instead of hasFeature(), as the name of the
constant implies.

Also detects support for high capacity, which some Newton camera supports,
and explictily disable it. It should be the default, but better safe than sorry.

Also update some comments about need to handle gain/readout dependency.