driver phenom: rely on the metadata to know the BPP in an image
Instead of relying on the requested BPP, use the value passed with the data.
In most cases it's the same. In practice, we never use BPP == 16, but
we've seen cases where just after the transition between BPP, it would
return the data of a different format than expected. Better be defensive.