gui acqmng: don't allow e-beam scale < 1, even on HQ preset
HQ preset divides the scale by 2 (to increase the resolution by 2).
However, recently the phenom driver added support for scale < 1,
to access lower FoV without chaning the horizontalFoV. This shouldn't
be used for normal use. Especially, the resolution stays 2048² anyway.
So the HQ preset was picking a scale of 0.5, if the current scale was 1.

=> Force at least a scale of 1.