gui comp: Fix validating values when focus changes on combo-box
The easy way of commit 71ebf5c12401 didn't work properly.
The main issue is that every time the drop-down hides/shows a KILL_FOCUS
events happens, which forces the value to be reset.

So do more like the text controls, and convert the (right) focus events
into a TEXT_ENTER event.
The trick is to not emit too many of the events, to avoid too many useless
requests to change the HFW.