driver phenom: don't send non-changes of door state
On every reception of the DOOR-STATUS event coming from the Phenom,
the driver would notify a change of .opened VA, even if the value
is still the same.

The new Phenom firmware v4.5 tends to generate a loooot of such events
without much reason (~ twice a second when the stage is moving).
Per se, it's not especially wrong, as the event (and the VA) indicates
that the door is still closed.
However some code in the GUI rely on the fact that VAs only notify
when the value changes. This should be fixed as well, but anyway,
it saves resources.