gui cont tabs: fix SEM magnification not working in SECOM alignment tab
After the refactoring of the SECOM alignment tab, the SEM VAs were set as local.
That means they work... iff the stream is playing. That's not the case if
spot mode is active, and they also don't get the value back when the VAs
are modified in other ways (eg, from the stream tab).

=> Make all the SEM VAs displayed as HW VAs, so they share the value globally.