driver phenom: simplify blanker management
Move all the code of the blanker on the scanner, which is where the VA is.
Remove support for old Phenom firmware, as all known DELPHI has the new firmware,
and if they don't they could just update.

Also add support for manually forcing the blanker. Not really needed but it's
really straightforward.