cli: accept None as attribute value to indicate None
Because we rely on YAML to do the translation from the command line to the raw value,
some values must be written differently than what we internally write.
For instance, floats with exponent _must_ have a period.
One other discreprancy is that None must be written "null". If you try to set None, odemis
will say something like "None is incorrect, it should be one of True, False, None".
=> Detect such case and convert the "None" string (which is left as-is by YAML) into
a (Python) None.