gui cont tabs: Also disable spot mode when leaving SPARC acquisition tab
Commit 12bbd4c4fb prevented streams previously playing when leaving the tab to
play again when entering the tab again.

That's good for normal streams, but for the spot stream, it's associated to the
spot mode (in the toolbar). When coming back to the tab, there was a discrepency
between the spot mode and the spot stream not being active. There are two ways to
solve this: either resume the spot stream if spot mode is active (but it's strange/
useless to have the spot mode active if the other stream is paused), or disable the
spot mode when leaving.

=> implement that second option. In particular, this fixes the spot mode not actually
starting when starting an optical stream.