gui comp canvas: remove world coordinates, and only use phyiscal coordinates
For historical reasons, we had two very similar coordinate systems: world and physical.
Eventually, world and physical became identical excepted that Y was pointing oppositely.

=> Simplify by only using Physical, and directly converting between physical <> buffer/view.

Note that this highlights one weird concequence of the convention of having the
physical Y opposite of the view Y. When displaying an image, the position in Y is
therefore multiplied by a negative scale, but the complete image is not inverted.
That means the first line of an image is at the smallest Y value, which can be
easily confusing.