driver semcomedi: avoid scanning margin pixels too long when settle time would be much smaller than dwell time
Until now the number of times the first pixel (margin) would be scanned would be
based on the settle time, rounded up. In most cases that makes complete sense,
as we better be "safe than sorry".
However, when the dwell time is very long (ie, several orders of magnitude) compared
to the settle time, that means the first pixel will be scanned twice. That could
delay quite a bit small resolution scans, and cause quite more damage to that first pixel,
without actually really improving the quality of the signal.