gui comp text: Use a unit prefix for 0, when this makes sense
In general, when 0 is displayed, the unit prefix doesn't make any sense (0 m = 0 nm).
However, in the GUI, it's nice to easily change values by typing them,
and this is done by automatically _not_ focusing the unit part.
So, for example when changing the wavelength, the user can easily endup
typing "0 m" -> "500 m" instead of "500 nm".

In other words, it's handy to show 0 with a "good" prefix.
=> We define "good" as "the prefix that would be used to show a small
value, which we find out by using key_step.