dataio tiff: be more robust in case saving a TIFF tag fails
Of course, normally saving the tags should work. However, if for some
reason it doesn't, it's still better to save the data than entirely fail.

This was encountered when saving TIFF files in parallel (via multiple threads).
The root cause was an issue in the pylibtiff, which has been fixed.
Still, it's good to keep the code robust for other potential problems.