driver omicronxx: ensure the emission status led is off when lights are off
When all the lights were off, the emission status led on the front panel of the LedHub was still on.
In practice, although no emission was going on, this was confusing for the user.
Turning off the master light when all the light sources are off ensures this status led
properly represents the reality.
This is done via the "stand-by" bits, as recommended by Omicron.

Also reorganise the code a bit to more properly support sub-devices,
and read temperature.