acq stream: fix getBoundingBox() to work or properly fail in all cases
The changes for displaying tiles broke live view.
The main required fix was to compute the bounding box in more cases, and in the cases
it cannot be done, properly report it, and properly handle it in the miccanvas.
Also fix displaying optical images, for which the first image had the metadata not
correctly updated.