acq stream: Make AR and Spectrum StaticStreams able to read DataArrayShadows
Now that the dataio.tiff converter supports open_data(), this is the default
method used by util.dataio to open TIFF files. This means even AR and Spectrum
streams, which currently have no chance to be pyramidal, will be opened
via a DAS. Per se, this is fine... but the streams need to handle it.

=> Add very basic support for DAS in the Streams, by calling .getData() ASAP.

Also extend test cases to check for this (and fix a couple of small issues in the test cases)