gui: fix opening a file without absolute path in absolute
Such command would fail:
$ odemis-gui --standalone --file 20170322-084203.ome.tiff

The main reason is that the current path was changed at init, before the file is opened.
So the data converter couldn't find the path.
In addition, as there was no path, the GUI widget displaying the path would get very sad.

=> Don't change the current path at init. Maybe it was necessary at some point,
but it's not recommended practice to depend on this, and currently it works when
staying in the current path.

=> If the file name is relative, change it to absolute before displaying in the GUI,
so that there is always a directory to show.