acq path: set the filter to "pass-through" for alignment on SPARCv2
It was done for SPARCv1, but not for SPARCv2.
Usually, the pass-through filter is on band 1, which is the default band
when starting the SR193, so it wasn't noticed often, but could cause issues
sometimes for the alignment.

There are still issues with the external spectrometer with a filter.
Need to make the code a lot smarter with such configuration, to find
out which detector is used.