gui conf util: fix settings.changed event being fired at the wrong time on sliders
The event should be sent when the wxEvents are used to update the VA.
On the sliders, there are options to which wxEvent is used to update the VA,
so the update.changed event trigger should depend on that option too.

This fixes estimated time not updating in the acquisition window when changing
the dwell time.