dataio tiff: don't use libtiff.tiff.TIFFfile in test case, but just the ctype wrapper
The python version of libtiff (TIFFfile) is not always working properly.
The latest version returns the Tags with string type with the \0 which
causes issues with the test case => just use the ctype wrapper, which
we are used to fix if there is an issue.

Also do some minor clean-ups in tiff.py.