acq path: add SECOM/DELPHI path managment to turn off/on the fan
The fan used to be always off when doing a high-quality acquisition.
This kind-of works, but if the exposure time of the fluorescence is long,
it might cause the CCD to acquire image with extra noise.
In addition, the recent change to increase the target temperature (in order
to make sure the fan doesn't turn on again after a while) caused the
temperature to raise faster on some cameras.

=> Do the "Right Thing" and only turn off the CCD fan when acquiring Streams
which don't use the CCD.

In order to do so, we need to know _when_ each stream start and stop.
That's made possible by using the OPM also for the SECOM/DELPHI.
It's not exactly setting the "optical path", but it takes care of
setting the hardware in the right settings for a given acquisition type.

We also need to introduce the notion of "acquisition quality" in order
to not turn off the fan during live imaging.