gui main: ensure the log button is the right colour at init
Since call_in_wx_main() can call directly the function if the call is
done in the main thread, the order of execution doesn't always follow
the order of calls.
What could happen is:
 * level = 20 (in sub thread)
 * level = 40 (in main thread)

It could result in this order of execution:
 * set button to 40
 * set button to 20
=> reset to grey

To workaround this, we always read the log level directly from the model (in the VA).