acq drift: make it clearer that .estimate() return the latest drift
It was talking of "orig_drift", which could easily be confused as the
total drift (since the origin). It actually only meant as the drift
when measured using the original image (but as we are acquiring using
the latest known drift.

=> Rename it to just ".drift", and also introduce ".tot_drift" which
represent the total drift, and can be useful in many cases.