gui cont tabs SECOM align: disable SEM settings during spot mode
During spot mode, the scanning settings are overridden to scan as a spot.
This was reflected on the SEM settings. Not only is it a bit strange for
the user, but if the user was changing it during the spot mode, it could
end up with lots of confusion and not even put back the proper settings
afterwards.
=> Disable the stream settings during spot mode.
In theory, we could still allow the user to change settings like
accel voltage or spot size, but let's keep it simple.