config: Make odemis-stop less willing to kill the backend
It was waiting only 1s per component before trying to kill them all.

In some cases, it can genuinely take a couple of seconds to stop a component.
In such a case it could make the hardware component in a bad state as the driver
was stopped half-way during shut-down. That was especially noticeable on some Andor cameras.

=> give 5 s per component