odemisd: handle better bad endings
If the backend receives a TERM signal, it will terminate.
If it's already doing so (via the standard call), it's no need to call terminate() again,
as it just tends to make things worse (calling terminate() on all the components twice,
sometimes simultaneously).

Also don't be too scared if the metadata unsubscribe fails, it likely just
because the component got killed.