gui cont streams: fix intensity range being set outside of data range
The intensity range value and its range were set in different functions.
As there were running asynchrnously (in the main GUI thread), it could
happen that the range was set before the new value was updated.
This lead to some cases where an old value of the intensity was set
while the new range would not allow it.

=> Move everything in the same function, and ensure the range is always
set first, then the latest known value is set.