gui comp buttons: Make sure not to hold the mouse capture when the button is disabled
It would be possible to trigger cases where a button keeps the mouse capture forever,
preventing the user from clicking anywhere else on the screen. Getting away from this
was only possible by quitting Odemis by pressing Alt+F4.

For instance, on a DELPHI, on the stream tab:
 1. Turn on the SEM stream
 2. Turn off the SEM stream
 3. Press down the SEM button with the mouse
 4. Press F6 (the SEM stream starts, the button gets disabled)
 5. Release the mouse button
 => It's captured

To avoid such scenario, we "patch" the GenButton class by overriding the Enable() method.
Enable() now automatically drops the capture when the button is disabled.