acq path: ensures that for the modes fiber-align and spec-fiber-focus all the axes are properly set
The recent change on automatically removing moves which don't affect the detector
caused issues for the modes fiber-align and spec-fiber-focus because they
are "special" as they don't really have a detector as "detector" component,
but a middle actuator.
Because of this, some moves in the mode specification were wrongly removed.

=> Use the .affects of the "detector" to get the real detector, and
convince the pruning code to keep all the moves.