driver semcomedi: ensure the computed scanner metadata has priority
The metadata is computed as the merge of the static metadata set on the scanner/main,
the detector, and some metadata computed on the fly (eg, POS, ACQ_DATE).
That last type of metadata used to be overriden by the detector. However,
it's computed on the fly for a reason: it's the only way to have it
proper, and so obviously it should have the highest priority.

=> Fixes some issues on SPARCs with a stage, which has non-properly aligned
data.