acq align: improve scanning size grid when SEM and CCD ratios are not 1:1
The previous code estimated the scanning size by using the SEM and CCD FoV _areas_.
That worked fine if the ratio is 1:1. However, what really matters is
the intersection of the FoVs, as if the two ratios are similar, the
scanned grid can be bigger, and if the two ratios are very different,
the scanned grid should be smaller.

=> Compute the scanned grid by comparing each dimension independently,
and pick the smallest dimension.