acq stream: fix setting drift correction region out of bound sometimes causes GUI error
The code was just plenly wrong. In case the DC region was out of bound, it would
try to change a tuple (which cannot be), or even change an item out of index in the tuple.
it was not usually happening because the GUI already checks the ROI is within bounds.

=> Use the same code as for the RoA.