acq align: move SEM image calibration just after the hole detection and autofocus
Now that SEM image calibration doesn't need optical image, it's better
to do it early so that the CL spot for alignment will more precisely placed.

There is also the potential to save time by staying near the hole (although this
is not happening now, at it goes in between to the bottom hole).