driver semcomedi: fix wrong range conversion of AO for scanner
The _array_from_phys() function expects the channels to be different
on the _last_ dimension, but _update_raw_scan_array() was placing them
on the first dimension.

As long as the channels use the same range, that doesn't matter.
But things could go wrong if the two channels need a different range.
In practice almost all hardware need a range < 5V, and with the NI card,
that would result in the same range all the time. However when needing
voltages > 5 V only on one axis, error would happen.