gui conf util: display filter bands properly when used in axis
On the SECOM, the filter bands were properly (eg, 565/40 nm) displayed,
but on the SPARC it used the raw tuple data. It also works, but it's ugly.
The reason was that the SPARC uses the filter as one of many axes, and so
that code was not aware of the notion of band.

=> Detect that the axis is a wavelength band, and in such case, use the right string conversion.