driver semcomedi: workaround 32 bit issues in comedi driver for AO period
When trying to find out the maximum period on the AO, on 32 bit systems
it seems the rounding method goes wrong (overflow) and ends up giving
the smallest period.

=> subtract a tiny bit more to be out of the overflow.
Using the minimum period as "tiny bit" should be good enough in every case.

We've tried to just use CMDF_ROUND_DOWN flag to avoid the overflow,
but it didn't seem to help.