gui model: make StreamView.getStreams() always return Streams
It would return directly stream_tree.getStreams(), but that means that now it
could return DataProjections. The rest of the methods of the View are hiding
the DataProjection, and this method should do too.

That simplifies some code (and maybe solve already some bugs).

If DataProjections are needed, then the caller should now explicitely call
View.stream_tree.getStreams().