acq mng: adjust the metadata also on partial acquisition failures
If the acquisition partially failed, the metadata would not be adjusted.
As most the "adjustment" is to merge the overlay stream, which is the last
stream acquired (the most likely one to fail), it wouldn't cause too much
loss. Still, the names of the streams wouldn't be carried in the metadata,
which is unconvenient.

=> Fix the name of the streams missing after partial acquisition failure.