util piconfig: automatically retry writting parameters
Some parameters sometimes don't work on the first time because other parameters
indicate they would be out of range. We could try to be super clever and
guess the right order. But it's just simpler to retry a couple of times.

We used to just indicate it in the doc, but it's even better to automatically do it.