gui comp stream: fix streambar not resizing properly after removing a stream panel
The previous change removed the use of CallAfter to fit_stream().
It turns out it _was_ useful. It helped to ensure the sizer had updated
its list of children when refitting the stream panels.

We could just put back CallAfter, but let's try to be more clever, and
explictily tell the sizer. This way, it's still possible to immediately
refit the stream. (=> no delay)