model vattributes: detect value changes when a VA has both getter and setter
VAs with both getter and setter can in theory have a different value every
time you read it.
That doesn't mean that it's not worthy to be notify when the value change.

Subscribers to such VA can expect that every call to the setter with a different
value cause a notification. This already works. When a value is read via the
getter, and a different value is set later, it should also cause a notification.
This wasn't working, and is fixed.

Note that still when reading a new value, no notfication is sent. This might need
to be changed too...